# docs/swagger.yaml
basePath: /
definitions:
  dto.RequestOTPRequest:
    properties:
      phone:
        type: string
    type: object
  dto.UserResponse:
    properties:
      phone:
        example: "09123456789"
        type: string
      registered_at:
        example: "2025-08-25T12:00:00Z"
        type: string
    type: object
  dto.VerifyOTPRequest:
    properties:
      otp:
        type: string
      phone:
        type: string
    type: object
  dto.VerifyOTPResponse:
    properties:
      token:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      message:
        example: Invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.PaginatedResponse-dto_UserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      message:
        type: string
      page:
        type: integer
      size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
  response.Response-any:
    properties:
      data: {}
      message:
        example: OK
        type: string
      success:
        example: true
        type: boolean
    type: object
  response.Response-dto_UserResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserResponse'
      message:
        example: OK
        type: string
      success:
        example: true
        type: boolean
    type: object
  response.Response-dto_VerifyOTPResponse:
    properties:
      data:
        $ref: '#/definitions/dto.VerifyOTPResponse'
      message:
        example: OK
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: OTP-based authentication and user management API.
  title: Dekamond Task API
  version: "1.0"
paths:
  /auth/request-otp:
    post:
      consumes:
      - application/json
      description: Generate OTP for the given phone (Iranian format).
      parameters:
      - description: Phone number (09XXXXXXXXX)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RequestOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/response.Response-any'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Request OTP
      tags:
      - Auth
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Validates OTP, registers user if new, and returns JWT.
      parameters:
      - description: Phone and OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.Response-dto_VerifyOTPResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verify OTP and login/register
      tags:
      - Auth
  /users:
    get:
      consumes:
      - application/json
      description: List users with optional search, pagination (requires JWT).
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      - description: Search by phone substring
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of users
          schema:
            $ref: '#/definitions/response.PaginatedResponse-dto_UserResponse'
        "400":
          description: Invalid query
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
  /users/{phone}:
    get:
      consumes:
      - application/json
      description: Retrieve a single user by phone number.
      parameters:
      - description: User phone
        in: path
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/response.Response-dto_UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by phone
      tags:
      - Users
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"